baltic/requirements.txt
aiodns==3.0.0
aiofile==3.8.1
aiohttp==3.8.3
aiosignal==1.2.0
aniso8601==7.0.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.7.1
Automat==20.2.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
caio==0.9.8
certifi==2021.10.8
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.0.12
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==4.0.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.13.0
django-csp==3.7
django-debug-toolbar==3.7.0
django-feature-policy==4.0.0
django-permissions-policy==4.13.0
django-phonenumber-field==7.0.0
django-referrer-policy==1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
frozenlist==1.3.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
multidict==6.0.2
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook-shim==0.1.0
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==304
pywinpty==2.0.6
pyzmq==23.2.0
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
Send2Trash==1.8.0
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
stream-chat==4.5.0
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
wrapt==1.14.1
yarl==1.8.1
zope.interface==5.5.0
